Added display data for sets and maps.

+++ app/src/main/java/RunTimeComplexity/ListRunTime.java
@@ -7,7 +7,7 @@ import java.util.Random;
 public class ListRunTime {
 	private static final Random RAND = new Random();
 	
-	public static float accessTime(List<Integer> list, int n) {
+	public static long accessTime(List<Integer> list, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			list.get(RAND.nextInt(n));
@@ -15,7 +15,7 @@ public class ListRunTime {
 		return Watch.getTime();
 	}
 	
-	public static float searchTime(List<Integer> list, int n) {
+	public static long searchTime(List<Integer> list, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			list.contains(RAND.nextInt(n));
@@ -23,7 +23,7 @@ public class ListRunTime {
 		return Watch.getTime();
 	}
 	
-	public static float randomInsertionTime(List<Integer> list, int n) {
+	public static long randomInsertionTime(List<Integer> list, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			list.add(RAND.nextInt(n), i);
@@ -31,7 +31,7 @@ public class ListRunTime {
 		return Watch.getTime();
 	}
 	
-	public static float sequentialInsertionTime(List<Integer> list, int n) {
+	public static long sequentialInsertionTime(List<Integer> list, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			list.add(i);
@@ -39,7 +39,7 @@ public class ListRunTime {
 		return Watch.getTime();
 	}
 	
-	public static float randomDeletionTime(List<Integer> list, int n) {
+	public static long randomDeletionTime(List<Integer> list, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			list.remove(RAND.nextInt(n - i));
@@ -47,7 +47,7 @@ public class ListRunTime {
 		return Watch.getTime();
 	}
 	
-	public static float sequentialDeletionTime(List<Integer> list, int n) {
+	public static long sequentialDeletionTime(List<Integer> list, int n) {
 		Watch.start();
 		for (int i = n - 1; i >= 0; i--) {
 			list.remove(i);
