+++ app/src/main/java/RunTimeComplexity/MapRunTime.java
@@ -6,7 +6,7 @@ import java.util.Random;
 public class MapRunTime {
 	private static final Random RAND = new Random();
 	
-	public static float accessTime(Map<Integer, Integer> map, int n) {
+	public static long accessTime(Map<Integer, Integer> map, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			map.get(RAND.nextInt(n));
@@ -14,7 +14,7 @@ public class MapRunTime {
 		return Watch.getTime();
 	}
 	
-	public static float searchTime(Map<Integer, Integer> table, int n) {
+	public static long searchTime(Map<Integer, Integer> table, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			table.containsKey(RAND.nextInt(n));
@@ -22,7 +22,7 @@ public class MapRunTime {
 		return Watch.getTime();
 	}
 	
-	public static float insertionTime(Map<Integer, Integer> table, int n) {
+	public static long insertionTime(Map<Integer, Integer> table, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			table.put(RAND.nextInt(n), i);
@@ -30,7 +30,7 @@ public class MapRunTime {
 		return Watch.getTime();
 	}
 	
-	public static float deletionTime(Map<Integer, Integer> table, int n) {
+	public static long deletionTime(Map<Integer, Integer> table, int n) {
 		Watch.start();
 		for (int i = 0; i < n; i++) {
 			table.remove(RAND.nextInt(n - i));
